Moved calculate_diff methods to gradients.py.
