+++ landlab/grid/grid_funcs.py
@@ -61,58 +61,6 @@ def resolve_values_on_links(grid, link_values):
                      grid.link_length), link_values))
 
 
-def calculate_diff_at_active_links(grid, node_values, out=None):
-    """Calculate differences of node values over active links.
-
-    Calculates the difference in quantity *node_values* at each active link
-    in the grid.
-
-    Parameters
-    ----------
-    grid : ModelGrid
-        A ModelGrid.
-    node_values : ndarray
-        Values at grid nodes.
-    out : ndarray, optional
-        Buffer to hold the result.
-
-    Returns
-    -------
-    ndarray
-        Differences across active links.
-    """
-    if out is None:
-        out = grid.empty(centering='active_link')
-    return np.subtract(node_values[grid.activelink_tonode],
-                       node_values[grid.activelink_fromnode], out=out)
-
-
-def calculate_diff_at_links(grid, node_values, out=None):
-    """Calculate differences of node values over links.
-
-    Calculates the difference in quantity *node_values* at each link in the
-    grid.
-
-    Parameters
-    ----------
-    grid : ModelGrid
-        A ModelGrid.
-    node_values : ndarray
-        Values at grid nodes.
-    out : ndarray, optional
-        Buffer to hold the result.
-
-    Returns
-    -------
-    ndarray
-        Differences across links.
-    """
-    if out is None:
-        out = grid.empty(centering='link')
-    return np.subtract(node_values[grid.node_at_link_head],
-                       node_values[grid.node_at_link_tail], out=out)
-
-
 def calculate_flux_divergence_at_nodes(grid, active_link_flux, out=None):
     """Calculate flux divergence at grid nodes.
